Splitting Athena and S3 tests.
