+++ tox.ini
@@ -9,7 +9,7 @@ deps =
        pytest-timeout
        moto
 commands =
-       pytest --timeout=600 -n 8 tests
+       pytest --timeout=900 -n 8 tests
 
 [testenv:py36]
 passenv = AWS_PROFILE AWS_DEFAULT_REGION AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY
@@ -17,4 +17,4 @@ deps =
        {[testenv]deps}
        pytest-cov
 commands =
-       pytest --timeout=600 --cov=awswrangler -n 8 tests
+       pytest --timeout=900 --cov=awswrangler -n 8 tests
